requirements.txt
psycopg2-binary
Django
djangorestframework
django-cors-headers
python-dotenv
djangorestframework-simplejwt
PyJWT
asgiref
sqlparse
pytz